Update pizzaOrder.py
attempted to write the additional input statements on my own, including converting the num_toppings input from str to int.  At first, I had used str, but then I remembered that the input will already be a string and we need to convert it to input for the toppings loop.